Cleaned home page, fixed setting app access
